requirements.txt
pandas==1.2.3
numpy==1.20.1
torch==1.7.0
albumentations==0.5.2
opencv_python_headless==4.5.1.48
opencv-python==4.5.1.48
torchvision==0.8.1
pycocotools==2.0.2
Pillow==9.1.0
